requirements.txt
kafka_python==2.0.2
nltk==3.7
praw==7.7.0
pyspark==3.1.1
python-dotenv==1.0.0
Requests==2.28.2
spacy==3.5.2
